requirements.txt
Fabric
Flask-SQLAlchemy
Flask-WTF
WTForms
coverage
flask-login
Flask-Mail
flask-alembic
werkzeug==0.9